+++ letsdance/core/crypto.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import re
+
 from cryptography.exceptions import InvalidSignature
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric.ed25519 import (
@@ -46,10 +48,27 @@ def verify_signature(signature: str, key: str, data: bytes) -> bool:
         return True
 
 
+public_key_pattern = re.compile(r"83e(0[1-9]|1[0-2])(\d\d)$")
+
+
 def get_valid_public_key_endings() -> tuple[str, str]:
     now = timezone.now()
     return f"ed{now.year}", f"ed{now.year + 1}"
 
 
 def validate_public_key(key: str) -> bool:
-    return key.endswith(get_valid_public_key_endings())
+    match = public_key_pattern.search(key)
+    if not match:
+        return False
+
+    month_str, year_str = match.groups()
+
+    # Some weird date logic since we only need month granularity, so I'm trying
+    # to keep this as efficient as possible for faster key generation.
+    max_months = (2000 + int(year_str)) * 12 + int(month_str)
+    min_months = max_months - 24
+
+    now = timezone.now().date()
+    now_months = now.year * 12 + now.month
+
+    return min_months <= now_months <= max_months